# api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/api/v1/gateway/stats": {
      "get": {
        "description": "Messages sent, messages received, devices, and API keys across your whole account.",
        "operationId": "GatewayController_getStats_v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Account totals.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GatewayStatsResponseDTO"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid, or revoked API key."
          }
        },
        "security": [
          {
            "x-api-key": []
          }
        ],
        "summary": "Get account totals",
        "tags": [
          "gateway"
        ]
      }
    },
    "/api/v1/gateway/devices": {
      "get": {
        "description": "Every phone paired with your account. The push token and hardware serial are never returned.",
        "operationId": "GatewayController_getDevices_v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Your devices.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeviceListResponseDTO"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid, or revoked API key."
          }
        },
        "security": [
          {
            "x-api-key": []
          }
        ],
        "summary": "List your devices",
        "tags": [
          "gateway"
        ]
      }
    },
    "/api/v1/gateway/devices/{id}": {
      "get": {
        "description": "Full state of one device, including its last heartbeat, battery, and SIM list.",
        "operationId": "GatewayController_getDevice_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Device id, from GET /gateway/devices.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The device.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeviceResponseDTO"
                }
              }
            }
          },
          "400": {
            "description": "The device id is not a valid id."
          },
          "401": {
            "description": "Missing, invalid, or revoked API key."
          },
          "404": {
            "description": "No device with that id on your account."
          }
        },
        "security": [
          {
            "x-api-key": []
          }
        ],
        "summary": "Get a device",
        "tags": [
          "gateway"
        ]
      }
    },
    "/api/v1/gateway/send-sms": {
      "post": {
        "description": "Sends one message to one or more recipients from a phone on your account. deviceId is optional: without it textbee uses your default device, otherwise the enabled device with the most recent heartbeat. Every recipient counts as one message against your plan.",
        "operationId": "GatewayController_sendSMSDeviceless_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendSMSRequestDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The send was accepted. Use smsBatchId to follow delivery. Acceptance is not delivery: the phone still has to be online.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SendSMSResponseDTO"
                }
              }
            }
          },
          "400": {
            "description": "No enabled device to send from, the device is disabled, your email is not verified, or the message could not be pushed to the phone."
          },
          "401": {
            "description": "Missing, invalid, or revoked API key."
          },
          "429": {
            "description": "Your daily, monthly, or per-batch plan limit is used up. The body says which one."
          }
        },
        "security": [
          {
            "x-api-key": []
          }
        ],
        "summary": "Send an SMS",
        "tags": [
          "gateway"
        ]
      }
    },
    "/api/v1/gateway/send-bulk-sms": {
      "post": {
        "description": "Sends a batch where every entry has its own text and recipients. deviceId is optional and resolves the same way as POST /gateway/send-sms.",
        "operationId": "GatewayController_sendBulkSMSDeviceless_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SendBulkSMSRequestDTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The batch was accepted.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SendSMSResponseDTO"
                }
              }
            }
          },
          "400": {
            "description": "No enabled device to send from, the device is disabled, your email is not verified, or the batch could not be pushed to the phone."
          },
          "401": {
            "description": "Missing, invalid, or revoked API key."
          },
          "429": {
            "description": "Your daily, monthly, or per-batch plan limit is used up. The body says which one."
          }
        },
        "security": [
          {
            "x-api-key": []
          }
        ],
        "summary": "Send several SMS in one call",
        "tags": [
          "gateway"
        ]
      }
    },
...
    "/api/v1/gateway/devices/{id}/sms/{smsId}": {
      "get": {
        "description": "A single sent or received message with its delivery timestamps. Use it to check the outcome of one recipient.",
        "operationId": "GatewayController_getSMSById_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Device id, from GET /gateway/devices.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "smsId",
            "required": true,
            "in": "path",
            "description": "Message id, from the message history response.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The message.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SMSResponseDTO"
                }
              }
            }
          },
          "400": {
            "description": "The device id is not a valid id."
          },
          "401": {
            "description": "Missing, invalid, or revoked API key."
          },
          "404": {
            "description": "No such message on this device."
          }
        },
        "security": [
          {
            "x-api-key": []
          }
        ],
        "summary": "Get one message",
        "tags": [
          "gateway"
        ]
      }
    },
    "/api/v1/gateway/devices/{id}/sms-batch/{smsBatchId}": {
      "get": {
        "description": "The batch returned by a send, plus one message per recipient. Poll this to see how a send is progressing.",
        "operationId": "GatewayController_getSmsBatchById_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Device id, from GET /gateway/devices.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "smsBatchId",
            "required": true,
            "in": "path",
            "description": "Batch id, returned as smsBatchId by the send endpoints.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The batch and its messages.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SMSBatchResponseDTO"
                }
              }
            }
          },
          "400": {
            "description": "The device id is not a valid id."
          },
          "401": {
            "description": "Missing, invalid, or revoked API key."
          },
          "404": {
            "description": "No such batch on this device."
          }
        },
        "security": [
          {
            "x-api-key": []
          }
        ],
        "summary": "Get a send batch",
        "tags": [
          "gateway"
        ]
      }
    },
...
  },
  "info": {
    "title": "textbee API",
    "description": "Send and receive SMS through an Android phone you own. Authenticate every request with an API key from your textbee dashboard, sent as the x-api-key header.",
    "version": "1.0",
    "contact": {},
    "license": {
      "name": "MIT",
      "url": "https://github.com/vernu/textbee/blob/main/LICENSE"
    }
  },
...
  "servers": [
    {
      "url": "https://api.textbee.dev"
    }
  ],
  "components": {
    "securitySchemes": {
      "x-api-key": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key",
        "description": "API key from your textbee dashboard, sent on every request."
      }
    },
    "schemas": {
...
      "GatewayStatsResponseDTO": {
        "type": "object",
        "properties": {
          "data": {
            "description": "Account totals.",
            "allOf": [
              {
                "$ref": "#/components/schemas/GatewayStatsDTO"
              }
            ]
          }
        },
        "required": [
          "data"
        ]
      },
...
      "DeviceResponseDTO": {
        "type": "object",
        "properties": {
          "data": {
            "description": "The device.",
            "allOf": [
              {
                "$ref": "#/components/schemas/DeviceDTO"
              }
            ]
          }
        },
        "required": [
          "data"
        ]
      },
      "DeviceListResponseDTO": {
        "type": "object",
        "properties": {
          "data": {
            "description": "Your devices.",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DeviceDTO"
            }
          }
        },
        "required": [
          "data"
        ]
      },
...
      "SendSMSResponseDTO": {
        "type": "object",
        "properties": {
          "data": {
            "description": "Outcome of the send.",
            "allOf": [
              {
                "$ref": "#/components/schemas/SendSMSResultDTO"
              }
            ]
          }
        },
        "required": [
          "data"
        ]
      },
...
      "SendBulkSMSRequestDTO": {
        "type": "object",
        "properties": {
          "messageTemplate": {
            "type": "string",
            "description": "Optional label for the batch. Each entry in messages carries its own text."
          },
          "messages": {
            "description": "Messages to send. Every message can target different recipients and use a different SIM.",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SMSData"
            }
          },
          "deviceId": {
            "type": "string",
            "description": "Device to send from. When omitted, uses your default device, or the enabled device with the most recent heartbeat."
          }
        },
        "required": [
          "messages"
        ]
      },
...
    }
  }
}